+++ src/main/java/com/noradltd/wumpus/CommandParser.java
@@ -47,6 +47,7 @@ class CommandParser {
                 addCommands(arg -> game.takeArrow(), "t", "take");
                 addCommands(arg -> game.vizualize(), "v", "viz"); // hidden
                 addCommands(arg -> game.saveState(), "r", "remember");
+                addCommands(arg -> game.requestLoad(), "f", "forget", "load");
             }
         });
     }